Move mobile v3 API type definitions into separate file
